Move order list state to a subtree of `orders`, in prep for adding an edit subtree.
