+++ client/extensions/woocommerce/state/ui/orders/reducer.js
@@ -2,45 +2,12 @@
  * External dependencies
  */
 import { combineReducers, keyedReducer } from 'state/utils';
-import { WOOCOMMERCE_UI_ORDERS_SET_QUERY } from 'woocommerce/state/action-types';
 
 /**
- * Tracks the current page of orders displayed for the current site.
- *
- * @param  {Object} state  Current state
- * @param  {Object} action Action payload
- * @return {Object}        Updated state
+ * Internal dependencies
  */
-export function currentPage( state = 1, action ) {
-	const { type, query } = action;
-	switch ( type ) {
-		case WOOCOMMERCE_UI_ORDERS_SET_QUERY:
-			return query && query.page ? query.page : state;
-		default:
-			return state;
-	}
-}
+import list from './list/reducer';
 
-/**
- * Tracks the current page of orders displayed for the current site.
- *
- * @param  {Object} state  Current state
- * @param  {Object} action Action payload
- * @return {Object}        Updated state
- */
-export function currentSearch( state = '', action ) {
-	const { type, query } = action;
-	switch ( type ) {
-		case WOOCOMMERCE_UI_ORDERS_SET_QUERY:
-			return query && ( 'undefined' !== typeof query.search ) ? query.search : state;
-		default:
-			return state;
-	}
-}
-
-const ordersReducer = combineReducers( {
-	currentPage,
-	currentSearch
-} );
-
-export default keyedReducer( 'siteId', ordersReducer );
+export default keyedReducer( 'siteId', combineReducers( {
+	list,
+} ) );
